+++ src/main/java/org/opentripplanner/transit/model/basic/TransitMode.java
@@ -2,11 +2,12 @@ package org.opentripplanner.transit.model.basic;
 
 import java.util.EnumSet;
 import java.util.Set;
+import org.opentripplanner.framework.doc.DocumentedEnum;
 
 /**
  * Equivalent to GTFS route_type or to NeTEx TransportMode.
  */
-public enum TransitMode {
+public enum TransitMode implements DocumentedEnum<TransitMode> {
   RAIL,
   COACH,
   SUBWAY,
@@ -51,5 +52,34 @@ public enum TransitMode {
 
   public boolean onStreet() {
     return ON_STREET_MODES.contains(this);
+  }
+
+  @Override
+  public String typeDescription() {
+    return "Routing modes for transit, including rail, bus, ferry, etc. Equivalent to [GTFS `route_type`](https://developers.google.com/transit/gtfs/reference/#routestxt) or to NeTEx TransportMode. ";
+  }
+
+  @Override
+  public String enumValueDescription() {
+    return switch (this) {
+      case RAIL -> "Used for intercity or long-distance travel.";
+      case COACH -> "Used for long-distance bus routes.";
+      case SUBWAY -> "Subway or Metro, used for any underground rail system within a metropolitan area.";
+      case BUS -> "Used for short- and long-distance bus routes.";
+      case TRAM -> "Tram, streetcar or light rail. Used for any light rail or street level system within a metropolitan area.";
+      case FERRY -> "Used for short- and long-distance boat service.";
+      case AIRPLANE -> "Taking an airplane";
+      case CABLE_CAR -> "Used for street-level cable cars where the cable runs beneath the car.";
+      case GONDOLA -> "Gondola or suspended cable car. Typically used for aerial cable cars where the car is suspended from the cable.";
+      case FUNICULAR -> "Used for any rail system that moves on steep inclines with a cable traction system.";
+      case TROLLEYBUS -> "Used for trolleybus systems which draw power from overhead wires using poles on the roof of the vehicle.";
+      case MONORAIL -> "Used for any rail system that runs on a single rail.";
+      case CARPOOL -> """
+       Private car trips shared with others.
+       
+       This is currently not specified in GTFS so we use the mode type values 1500-1560 which are in the range of private taxis.
+       """;
+      case TAXI -> "Using a taxi service";
+    };
   }
 }